+++ src/main/java/net/geforcemods/securitycraft/entity/sentry/Sentry.java
@@ -11,7 +11,7 @@ import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.api.Owner;
 import net.geforcemods.securitycraft.blockentities.DisguisableBlockEntity;
 import net.geforcemods.securitycraft.blocks.DisguisableBlock;
-import net.geforcemods.securitycraft.blocks.SentryDisguiseBlock;
+import net.geforcemods.securitycraft.blocks.SometimesVisibleBlock;
 import net.geforcemods.securitycraft.items.ModuleItem;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.network.client.InitSentryAnimation;
@@ -224,7 +224,7 @@ public class Sentry extends PathfinderMob implements RangedAttackMob, IEMPAffect
 					Block.popResource(level, pos, new ItemStack(SCContent.SPEED_MODULE.get()));
 
 				getSentryDisguiseBlockEntity().ifPresent(be -> be.removeModule(ModuleType.DISGUISE, false));
-				level.setBlockAndUpdate(blockPosition(), level.getBlockState(blockPosition()).setValue(SentryDisguiseBlock.INVISIBLE, true));
+				level.setBlockAndUpdate(blockPosition(), level.getBlockState(blockPosition()).setValue(SometimesVisibleBlock.INVISIBLE, true));
 				entityData.set(ALLOWLIST, new CompoundTag());
 				entityData.set(HAS_SPEED_MODULE, false);
 			}
@@ -422,7 +422,7 @@ public class Sentry extends PathfinderMob implements RangedAttackMob, IEMPAffect
 
 				if (!module.isEmpty() && module.getItem() instanceof ModuleItem moduleItem && moduleItem.getBlockAddon(module.getOrCreateTag()) != null) {
 					be.insertModule(module, false);
-					level.setBlockAndUpdate(blockPosition(), level.getBlockState(blockPosition()).setValue(SentryDisguiseBlock.INVISIBLE, false));
+					level.setBlockAndUpdate(blockPosition(), level.getBlockState(blockPosition()).setValue(SometimesVisibleBlock.INVISIBLE, false));
 				}
 			}
 		});
@@ -460,7 +460,7 @@ public class Sentry extends PathfinderMob implements RangedAttackMob, IEMPAffect
 				//remove a possibly existing old disguise module
 				be.removeModule(ModuleType.DISGUISE, false);
 				be.insertModule(module, false);
-				level.setBlockAndUpdate(blockPosition(), level.getBlockState(blockPosition()).setValue(SentryDisguiseBlock.INVISIBLE, false));
+				level.setBlockAndUpdate(blockPosition(), level.getBlockState(blockPosition()).setValue(SometimesVisibleBlock.INVISIBLE, false));
 			});
 		}
 	}