+++ src/main/java/net/geforcemods/securitycraft/screen/IMSScreen.java
@@ -6,13 +6,12 @@ import com.mojang.blaze3d.vertex.PoseStack;
 
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.blockentities.IMSBlockEntity;
-import net.geforcemods.securitycraft.blockentities.IMSBlockEntity.IMSTargetingMode;
+import net.geforcemods.securitycraft.misc.TargetingMode;
 import net.geforcemods.securitycraft.network.server.SyncIMSTargetingOption;
 import net.geforcemods.securitycraft.screen.components.ToggleComponentButton;
 import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.screens.Screen;
-import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
 
@@ -24,7 +23,7 @@ public class IMSScreen extends Screen {
 	private int leftPos;
 	private int topPos;
 	private IMSBlockEntity be;
-	private IMSTargetingMode targetMode;
+	private TargetingMode targetMode;
 
 	public IMSScreen(IMSBlockEntity be) {
 		super(be.getName());
@@ -38,7 +37,7 @@ public class IMSScreen extends Screen {
 
 		leftPos = (width - imageWidth) / 2;
 		topPos = (height - imageHeight) / 2;
-		addRenderableWidget(new ToggleComponentButton(width / 2 - 75, height / 2 - 38, 150, 20, this::updateButtonText, targetMode.ordinal(), 3, this::modeButtonClicked));
+		addRenderableWidget(new ToggleComponentButton(width / 2 - 75, height / 2 - 38, 150, 20, index -> TargetingMode.values()[index].translate(), targetMode.ordinal(), 3, this::modeButtonClicked));
 	}
 
 	@Override
@@ -68,12 +67,8 @@ public class IMSScreen extends Screen {
 	}
 
 	private void modeButtonClicked(Button button) {
-		targetMode = IMSTargetingMode.values()[((ToggleComponentButton) button).getCurrentIndex()];
+		targetMode = TargetingMode.values()[((ToggleComponentButton) button).getCurrentIndex()];
 		be.setTargetingMode(targetMode);
 		SecurityCraft.CHANNEL.sendToServer(new SyncIMSTargetingOption(be.getBlockPos(), be.getTargetingMode()));
-	}
-
-	private Component updateButtonText(int index) {
-		return Utils.localize("gui.securitycraft:srat.targets" + ((index + 2) % 3 + 1));
 	}
 }